backend/pyproject.toml
[project]
name = "weather-eval-backend"
version = "0.1.0"
description = "FastAPI backend for weather-eval"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.116.0",
  "google-cloud-firestore>=2.21.0",
  "pydantic-settings>=2.7.0",
  "uvicorn[standard]>=0.35.0",
]

[dependency-groups]
dev = [
  "httpx>=0.28.1",
  "pytest>=8.3.5",
  "ruff>=0.9.7",
]

[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "A", "C4", "SIM"]

[tool.pytest.ini_options]
pythonpath = ["."]
addopts = "-q"
